[AF0102] Finish challenge  and extra
Created an app with one activity MainActivity. With two buttons and a TextView.

One Button changes the value stored in the text view and the other shows a Toast message that shows the user the current value stored in the text view.

The buttons shares  a one onClick method the MainActOnClick. On the basis of the different button id I call the right function for the right button click. using the switch case command.

I added a landscape layout for the main activity and saved the value in the ovverriden onSaveInstanceState() method in the bundle wiht a created constant string as key,  to prevent loesing the value by rotating the screen of the application.